+++ fabric/src/main/java/net/merchantpug/bovinesandbuttercups/mixin/fabric/LivingEntityMixin.java
@@ -47,6 +47,7 @@ public abstract class LivingEntityMixin extends Entity {
                 }
             }));
             Services.COMPONENT.setLockdownMobEffects(entity, lockdownEffectsToUpdate);
+            Services.COMPONENT.syncLockdownMobEffects(entity);
         }
     }