+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Bugfixes
 - Fixed ImmutableMap crash upon consuming nectar.
 - Fixed pack formats for built-in resource packs.
+- Fixed syncing of Lockdown effects.